requirements.txt
opencv-python
mediapipe
numpy
PyYAML
fer